requirements.txt
flask
aiohttp
python-dotenv
brotli
beautifulsoup4